+++ concordium-consensus/src/Concordium/KonsensusV1/LeaderElection.hs
@@ -76,24 +76,41 @@ updateWithBlockNonce ::
     Hash.Hash
 updateWithBlockNonce bn un = Hash.hash $ runPut $ put un <> put (proofToHash bn)
 
--- |Update the running 'updatedNonce' in seed state.  Blocks after the trigger block do not
--- contribute.  If the timestamp is at least the epoch transition time, then the
+-- |Update the running 'updatedNonce' in seed state. Blocks after the trigger block do not
+-- contribute. If the timestamp is at least the epoch transition time, then the
 -- 'epochTransitionTriggered' flag is set in the seed state, indicating that a new epoch should
 -- begin once the block is finalized.
+--
+-- If the 'shutdownTriggered' flag is set in the seed state, then this indicates that
+-- a protocol update has become effective. Note however, that the protocol update itself
+-- will not take place before following the epoch transition.
 updateSeedStateForBlock ::
     -- |Timestamp of the block
     Timestamp ->
     -- |Block nonce of the block
     BlockNonce ->
+    -- |Protocol update effective
+    Bool ->
     -- |Prior seed state
     SeedState 'SeedStateVersion1 ->
     -- |Updated seed state
     SeedState 'SeedStateVersion1
-updateSeedStateForBlock ts bn ss
+updateSeedStateForBlock ts bn isEffective ss
+    -- Epoch tranisition is already in progress.
     | ss ^. epochTransitionTriggered = ss
-    | ss ^. triggerBlockTime <= ts = ss' & epochTransitionTriggered .~ True
+    -- The block timestamp is beyond the trigger time of the epoch
+    -- and there is also a protocol update effective.
+    | isBlockAfterTriggerTime,
+      isEffective =
+        ss'
+            & epochTransitionTriggered .~ True
+            & shutdownTriggered .~ True
+    -- The block timestamp is after the epoch trigger time,
+    -- so flag it.
+    | isBlockAfterTriggerTime = ss' & epochTransitionTriggered .~ True
     | otherwise = ss'
   where
+    isBlockAfterTriggerTime = ss ^. triggerBlockTime <= ts
     ss' = ss & updatedNonce %~ updateWithBlockNonce bn
 
 -- |Compute the leadership election nonce for the new epoch.
@@ -109,6 +126,8 @@ nonceForNewEpoch newBakers SeedStateV1{..} = Hash.hash $ runPut $ do
     putFullBakers newBakers
 
 -- |Update the seed state to account for a transition to a new epoch.
+-- Note: This function should never be called on a seed state with the `ss1ShutdownTriggered`
+-- flag set.
 updateSeedStateForEpoch ::
     -- |Bakers for the new epoch
     FullBakers ->
@@ -121,6 +140,7 @@ updateSeedStateForEpoch newBakers epochDuration ss =
     SeedStateV1
         { ss1Epoch = ss ^. epoch + 1,
           ss1EpochTransitionTriggered = False,
+          ss1ShutdownTriggered = False,
           ss1TriggerBlockTime = (ss ^. triggerBlockTime) `addDuration` epochDuration,
           ss1UpdatedNonce = newNonce,
           ss1CurrentLeadershipElectionNonce = newNonce